# Esercizio02Portale/esercizio02PortaleApi/src/main/resources/openapi.artista.yaml
openapi: 3.0.3
info:
  title: Artists API
  description: API for managing artists, with CRUD operations and advanced search capabilities.
  version: 1.0.0
  contact:
    email: faprile@studenti.apuliadigitalmaker.it

paths:
  /artists:
    post:
      tags:
        - artist
      summary: Create a new artist
      description: Add a new artist to the database
      operationId: createArtist
      requestBody:
        description: Data for the new artist
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artist'
      responses:
        '201':
          description: Artist created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Invalid input

    get:
      tags:
        - artist
      summary: List all artists
      description: Returns a list of all artists
      operationId: getAllArtists
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '404':
          description: No artists found

  /artists/{artistId}:
    get:
      tags:
        - artist
      summary: Retrieve a single artist
      description: Get the details of an artist by ID
      operationId: getArtistById
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          description: Artist not found

    put:
      tags:
        - artist
      summary: Update an artist
      description: Update the details of an existing artist
      operationId: updateArtist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated artist data
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artist'
      responses:
        '204':
          description: Artist updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          description: Artist not found

    delete:
      tags:
        - artist
      summary: Delete an artist
      description: Delete an artist by ID
      operationId: deleteArtist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Artist deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          description: Artist not found

  /artists/nationality/{nationality}:
    get:
      tags:
        - artist
      summary: List artists by nationality
      description: Get all artists of a specific nationality
      operationId: getArtistsByNationality
      parameters:
        - name: nationality
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '404':
          description: No artists found for this nationality

  /artists/year/{year}:
    get:
      tags:
        - artist
      summary: List artists by birth year
      description: Get all artists born in a specific year
      operationId: getArtistsByYear
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '404':
          description: No artists found for this year

components:
  schemas:
    Artist:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        type:
          type: string
          enum: [solo, group]
        name:
          type: string
        surname:
          type: string
        stageName:
          type: string
        birthYear:
          type: integer
        biography:
          type: string
        genre:
          type: string
          enum: [pop, rock, jazz]
        nationality:
          type: string
      required:
        - name
        - surname
        - stageName
        - birthYear
        - biography
        - genre
        - nationality